Move Search to its own file
