+++ main.py
@@ -1,10 +1,10 @@
-from flask import Flask, request, jsonify
+from flask import Flask, request
 from flask_restful import Resource, Api
 import sqlite3 as sql
 import traceback, sys
 
-DB_FILE = 'HeliumDB.db'
-SEARCH_PAGE_SIZE = 20
+import consts
+from search import Search
 
 app = Flask(__name__, static_url_path='/')
 api = Api(app)
@@ -14,49 +14,8 @@ def index():
 
 app.add_url_rule('/', 'index', index)
 
-with sql.connect(DB_FILE) as conn:
-    file = open('./resources/schema.sql', mode = 'r', encoding='utf-8')
-    c = conn.cursor()
-    c.executescript(file.read())
 
-
-class Quote(Resource):
-    def get(self):
-        args = request.args
-        with sql.connect(DB_FILE) as conn:
-            conn.row_factory = sql.Row
-            c = conn.cursor()
-            #return jsonify({"message": 200, "success":True, "data": [dict(x) for x in (c.execute('SELECT * FROM quote').fetchall())]})
-            return [dict(x) for x in (c.execute('SELECT * FROM quote').fetchall())]
-
-#TODO: Book availability
-class Search(Resource):
-
-    def get(self):
-        args = {'q':request.args.get('q', '').lower()}
-
-        with sql.connect(DB_FILE) as conn:
-            conn.row_factory = sql.Row
-            c = conn.cursor()
-
-            sql_query = '''
-            WITH q_authors AS (
-                SELECT GROUP_CONCAT(Name,', ') AS Author_names, Isbn
-                FROM AUTHORS NATURAL JOIN BOOK_AUTHORS
-                GROUP BY Isbn
-            )
-            SELECT Isbn, Title, Author_names, Cover_url
-            FROM BOOK NATURAL JOIN q_authors
-            WHERE INSTR(LOWER(Title), :q) > 0
-            OR INSTR(LOWER(Author_names), :q) > 0
-            OR INSTR(LOWER(Isbn), :q) > 0
-            ORDER BY INSTR(LOWER(Isbn), :q), INSTR(LOWER(Author_names), :q), INSTR(LOWER(Title), :q)
-            ;
-            '''
-
-            return [dict(x) for x in c.execute(sql_query, args).fetchmany(size=SEARCH_PAGE_SIZE)]
-
-
+#TODO: refactor as a Resource put method
 def create_borrower():
     args = request.args
 
@@ -80,11 +39,15 @@ def create_borrower():
         # return [dict(x) for x in (c.execute('SELECT * FROM quote').fetchall())]
         return app.make_response(200)
 
-
-api.add_resource(Quote, '/quote', endpoint='quote')
-api.add_resource(Search, '/search', endpoint='search')
-
 app.add_url_rule('/borrower/create', 'create_borrower', create_borrower, methods=["POST"])
 
+
+api.add_resource(Search, '/search', endpoint='search')
+
+
 if __name__ == '__main__':
+    with sql.connect(DB_FILE) as conn:
+        file = open('./resources/schema.sql', mode = 'r', encoding='utf-8')
+        c = conn.cursor()
+        c.executescript(file.read())
     app.run(debug=True)
